requirements.txt
asgiref==3.8.1
astroid==3.2.4
dill==0.3.8
Django==5.0.7
djangorestframework==3.15.2
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
packaging==24.1
platformdirs==4.2.2
pluggy==1.5.0
psycopg2-binary==2.9.9
pylint==3.2.6
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pytest==8.3.2
pytest-django==4.8.0
sqlparse==0.5.1
tomlkit==0.13.0